Update projects.json
Add `wallets` to the project schema. A type of `primary` is used to denote a wallet that is primarily used by the project, and a type of `authoritative` is to denote a wallet address that is or was authorized to create or own some assets (farms, tokens, etc) at some point.